+++ info/info.go
@@ -106,7 +106,7 @@ func getKeymaps(screenWidth int) string {
 	keys.AddSeparator()
 	keys.AddKeymap("Read comments", "Enter")
 	keys.AddKeymap("Change category", "Tab")
-	keys.AddKeymap("Open submission link in browser", "o")
+	keys.AddKeymap("Open story in browser", "o")
 	keys.AddKeymap("Open comments in browser", "c")
 	keys.AddKeymap("Refresh", "r")
 	keys.AddSeparator()